refactor: split store package by responsibility
